+++ src/netrace/netrace.h
@@ -95,7 +95,7 @@ struct nt_packet {
 	unsigned short dst;  // Changed from unsigned char to support >255 nodes
 	unsigned char node_types;
 	unsigned char num_deps;
-	unsigned short custom_size;  // Custom packet size in bytes (0-65535) - MUST be before deps pointer!
+	unsigned int custom_size;    // Custom packet size in bytes (0-4294967295) - MUST be before deps pointer!
 	nt_dependency_t* deps;       // Pointer field MUST be last to match binary format
 };
